agent/hardgate_agent/requirements.txt
# Core Dependencies
google-adk>=0.1.0
litellm>=1.0.0

# CodeGates Integration
requests>=2.28.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Security and Analysis
bandit>=1.7.0
safety>=2.3.0
semgrep>=1.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Reporting and Output
jinja2>=3.1.0
markdown>=3.4.0

# External Integrations
splunk-sdk>=1.7.0
appdynamics>=4.5.0
jira>=3.5.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional Dependencies
# Uncomment if using specific features
# kubernetes>=26.0.0
# docker>=6.0.0
# terraform>=1.0.0 